backend/requirements.txt
fastapi
openai
langchain
faiss-cpu
python-dotenv
uvicorn
pandas